requirements.txt
colorama
pystyle
selenium
requests
python-dotenv
openai
psutil
seleniumbase
pyinstaller
